% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_confint.R
\name{text_confint}
\alias{text_confint}
\alias{text_confint.default}
\alias{text_confint.lm}
\alias{text_confint.lavaan}
\title{A Confidence Interval as Text}
\usage{
text_confint(x, ...)

\method{text_confint}{default}(
  x,
  param = NULL,
  level = 0.95,
  digits = getOption("out2text_coef_digits", 3),
  format = getOption("out2text_coef_format", "f"),
  sep = getOption("out2text_confint_sep", ", "),
  brackets = getOption("out2text_confint_brackets", c("[", "]")),
  ...
)

\method{text_confint}{lm}(
  x,
  param = NULL,
  level = 0.95,
  digits = getOption("out2text_coef_digits", 3),
  format = getOption("out2text_coef_format", "f"),
  sep = getOption("out2text_confint_sep", ", "),
  brackets = getOption("out2text_confint_brackets", c("[", "]")),
  ...
)

\method{text_confint}{lavaan}(
  x,
  param = NULL,
  level = 0.95,
  digits = getOption("out2text_coef_digits", 3),
  format = getOption("out2text_coef_format", "f"),
  sep = getOption("out2text_confint_sep", ", "),
  brackets = getOption("out2text_confint_brackets", c("[", "]")),
  standardized = FALSE,
  group = 1L,
  ...
)
}
\arguments{
\item{x}{A supported object with coefficients in the results.}

\item{...}{Optional arguments. To be passed to
\code{\link[stats:confint]{stats::confint()}}.}

\item{param}{The name of the term or parameter for which
the coefficients will be returned.}

\item{level}{The level of confidence. Default is .95.}

\item{digits}{The number of decimal places to keep. To be
passed to \code{\link[=formatC]{formatC()}}.}

\item{format}{The format string used in \code{\link[=formatC]{formatC()}}. Default
is \code{"f"}.}

\item{sep}{The separator. Default is \code{", "}.}

\item{brackets}{The brackets. A character vector with
two elements. Default is \code{c("[", "]")}.}

\item{standardized}{Logical. Whether the estimate in the standardized
solution is to be returned. Default is \code{FALSE}.}

\item{group}{For an \code{lavaan}-class object, the group from which the
estimate is to be returned. Default is 1.}
}
\value{
A one-element character vector of the interval.
}
\description{
Extracts a confidence interval from an object
and returns the interval as a text string.
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{

set.seed(894252)
n <- 10
x <- rnorm(n)
y <- .6 * x + rnorm(n, 0, .8)
out <- lm(y ~ x)
text_confint(out, "x")

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
