% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_rsq.R
\name{text_rsq}
\alias{text_rsq}
\alias{text_rsq.default}
\alias{text_rsq.lm}
\title{A R-square as Text}
\usage{
text_rsq(x, ...)

\method{text_rsq}{default}(
  x,
  digits = getOption("out2text_coef_digits", 3),
  format = getOption("out2text_coef_format", "f"),
  ...
)

\method{text_rsq}{lm}(
  x,
  digits = getOption("out2text_coef_digits", 3),
  format = getOption("out2text_coef_format", "f"),
  type = c("unadjusted", "adjusted"),
  ...
)
}
\arguments{
\item{x}{A supported object with an R-square in the results.}

\item{...}{Optional arguments. Ignored for now.}

\item{digits}{The number of decimal places to keep. To be
passed to \code{\link[=formatC]{formatC()}}.}

\item{format}{The format string used in \code{\link[=formatC]{formatC()}}. Default
is \code{"f"}.}

\item{type}{The type of R-squared to be extracted.
Either \code{"unadjusted"}, the default, or \code{"adjusted"},
which returns the adjusted R-squared.}
}
\value{
A one-element character vector.
}
\description{
Extracts the R-square from an object
and returns it as a text string.
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{

set.seed(894252)
n <- 10
x <- rnorm(n)
y <- .6 * x + rnorm(n, 0, .8)
out <- lm(y ~ x)
text_rsq(out)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
