% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_coef.R
\name{text_coef}
\alias{text_coef}
\title{A Coefficient as Text}
\usage{
text_coef(x, ...)
}
\arguments{
\item{x}{A supported object with coefficients in the results.}

\item{...}{Optional arguments. Ignored for now.}

\item{param}{The name of the term or parameter for which
the coefficients will be returned.}

\item{digits}{The number of decimal places to keep. To be
passed to \code{\link[=formatC]{formatC()}}.}

\item{format}{The format string used in \code{\link[=formatC]{formatC()}}. Default
is \code{"f"}.}
}
\value{
A one-element character vector.
}
\description{
Extracts a coefficient from an object
and returns it as a text string.
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{

set.seed(894252)
n <- 10
x <- rnorm(n)
y <- .6 * x + rnorm(n, 0, .8)
out <- lm(y ~ x)
text_coef(out, "x")

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
