% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_confint.R
\name{text_confint}
\alias{text_confint}
\title{A Confidence Interval as Text}
\usage{
text_confint(x, ...)
}
\arguments{
\item{x}{A supported object with coefficients in the results.}

\item{...}{Optional arguments. To be passed to
\code{\link[stats:confint]{stats::confint()}}.}

\item{param}{The name of the term or parameter for which
the coefficients will be returned.}

\item{level}{The level of confidence. Default is .95.}

\item{digits}{The number of decimal places to keep. To be
passed to \code{\link[=formatC]{formatC()}}.}

\item{format}{The format string used in \code{\link[=formatC]{formatC()}}. Default
is \code{"f"}.}

\item{sep}{The separator. Default is \code{", "}.}

\item{brackets}{The brackets. A character vector with
two elements. Default is \code{c("[", "]")}.}
}
\value{
A one-element character vector of the interval.
}
\description{
Extracts a confidence interval from an object
and returns the interval as a text string.
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{

set.seed(894252)
n <- 10
x <- rnorm(n)
y <- .6 * x + rnorm(n, 0, .8)
out <- lm(y ~ x)
text_confint(out, "x")

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
